add a limit on the total number of history records.

+++ v4.0/src/renderer/components/History.tsx
@@ -18,8 +18,9 @@ import {
   Flex,
   HStack,
   List,
-  ListIcon,
   ListItem,
+  Select,
+  Spacer,
   Spinner,
   VStack,
 } from '@chakra-ui/react'
@@ -28,9 +29,9 @@ import { actions } from '@renderer/core/agent'
 import useOnBroadcast from '@renderer/core/useOnBroadcast'
 import { IHostsHistoryObject } from '@root/common/data'
 import dayjs from 'dayjs'
+import prettyBytes from 'pretty-bytes'
 import React, { useEffect, useRef, useState } from 'react'
 import { BiDetail, BiHistory, BiTrash } from 'react-icons/bi'
-import prettyBytes from 'pretty-bytes'
 
 interface IHistoryProps {
   list: IHostsHistoryObject[];
@@ -110,8 +111,10 @@ const History = () => {
   const [ is_loading, setIsLoading ] = useState(true)
   const [ list, setList ] = useState<IHostsHistoryObject[]>([])
   const [ selected_item, setSelectedItem ] = useState<IHostsHistoryObject>()
-  const { lang } = useModel('useI18n')
+  const [ history_limit, setHistoryLimit ] = useState(0)
   const btn_close = useRef(null)
+
+  const { lang } = useModel('useI18n')
 
   const loadData = async () => {
     // setIsLoading(true)
@@ -121,6 +124,10 @@ const History = () => {
     if (!selected_item) {
       setSelectedItem(list[0])
     }
+
+    let v = await actions.configGet('history_limit')
+    setHistoryLimit(v)
+
     setIsLoading(false)
 
     return list
@@ -132,6 +139,10 @@ const History = () => {
   }
 
   const deleteItem = async (id: string) => {
+    if (!confirm(lang.system_hosts_history_delete_confirm)) {
+      return
+    }
+
     let idx = list.findIndex(i => i.id === id)
     await actions.deleteHistory(id)
     setSelectedItem(undefined)
@@ -143,14 +154,29 @@ const History = () => {
     }
   }
 
+  const updateHistoryLimit = async (value: number) => {
+    if (!value || value < 0) return
+
+    setHistoryLimit(value)
+    await actions.configSet('history_limit', value)
+  }
+
   useEffect(() => {
-    loadData()
-  }, [])
+    if (is_open && list.length === 0) {
+      loadData()
+    }
+  }, [ is_open, list ])
 
   useOnBroadcast('show_history', () => {
     setIsOpen(true)
     loadData()
   })
+
+  let history_limit_values: number[] = [ 10, 50, 100, 500 ]
+  if (!history_limit_values.includes(history_limit)) {
+    history_limit_values.push(history_limit)
+    history_limit_values.sort()
+  }
 
   return (
     <Drawer
@@ -181,19 +207,33 @@ const History = () => {
             }
           </DrawerBody>
           <DrawerFooter>
-            <Button
-              leftIcon={<BiTrash/>}
-              variant="outline"
-              mr={3}
-              colorScheme="pink"
-              isDisabled={!selected_item}
-              onClick={() => selected_item && deleteItem(selected_item.id)}
-            >
-              {lang.delete}
-            </Button>
-            <Button onClick={onClose} ref={btn_close}>
-              {lang.close}
-            </Button>
+            <Flex width="100%" align="center">
+              <Box>{lang.system_hosts_history_limit}</Box>
+              <Box>
+                <Select
+                  value={history_limit}
+                  onChange={e => updateHistoryLimit(parseInt(e.target.value))}
+                >
+                  {history_limit_values.map(v => (
+                    <option key={v} value={v}>{v}</option>
+                  ))}
+                </Select>
+              </Box>
+              <Spacer/>
+              <Button
+                leftIcon={<BiTrash/>}
+                variant="outline"
+                mr={3}
+                colorScheme="pink"
+                isDisabled={!selected_item}
+                onClick={() => selected_item && deleteItem(selected_item.id)}
+              >
+                {lang.delete}
+              </Button>
+              <Button onClick={onClose} ref={btn_close}>
+                {lang.close}
+              </Button>
+            </Flex>
           </DrawerFooter>
         </DrawerContent>
       </DrawerOverlay>
